+++ examples/persistent_state/lib/app/domain/models/city.dart
@@ -6,10 +6,9 @@ part 'city.freezed.dart';
 part 'city.g.dart';
 
 @freezed
-sealed class City with _$City {
+class City with _$City {
   const factory City({
-    @JsonKey(name:'city')
-    required String name,
+    @JsonKey(name: 'city') required String name,
     required String state,
   }) = _City;